实现文本子项绘制方法 进行测试

+++ me.cqp.luohuaming.UnraidMonitor.PublicInfos/Drawing/Items/DrawingItem_Alert.cs
@@ -8,7 +8,7 @@ namespace me.cqp.luohuaming.UnraidMonitor.PublicInfos.Drawing.Items
 
         public override DrawingBase.Layout Layout { get; set; } = DrawingBase.Layout.Fill;
 
-        public override double FillPercentage { get; set; } = 100;
+        public override float FillPercentage { get; set; } = 100;
 
         public string Header { get; set; }
